+++ test/net/request.js
@@ -1,7 +1,7 @@
 'use strict';
 
-const { describe, test } = require('node:test')
-const assert = require('node:assert/strict')
+const { describe, test } = require('node:test');
+const assert = require('node:assert/strict');
 
 const { request } = require('../../src/net/request');
 const StunResponse = require('../../src/message/response');
@@ -11,7 +11,6 @@ const StunServer = require('../../src/net/dgram-server');
 const { createServer } = require('../../src/net/create-server');
 
 test('should work', (t, done) => {
-
   request('stun://stun.l.google.com:19302', (error, res) => {
     assert.equal(error, null);
     assert.equal(true, res instanceof StunResponse);
@@ -26,7 +25,6 @@ test('should work as promise', async () => {
 });
 
 test('url normalization should work', (t, done) => {
-
   request('stun.l.google.com:19302', (error, res) => {
     assert.equal(error, null);
     assert.equal(true, res instanceof StunResponse);
@@ -36,15 +34,16 @@ test('url normalization should work', (t, done) => {
 });
 
 test('should use provided STUN server', (t, done) => {
-
   const socket = {
     on: function () {},
     once: function () {},
     removeListener: function () {},
   };
   const server = new StunServer(socket);
-  let called = 0
-  server.send = function () { called++; }
+  let called = 0;
+  server.send = function () {
+    called++;
+  };
 
   request('stun.l.google.com:19302', { server, retries: 0 }, () => {
     assert.equal(called, 1);
@@ -53,7 +52,6 @@ test('should use provided STUN server', (t, done) => {
 });
 
 test('should use provided message', (t, done) => {
-
   const server = createServer({ type: 'udp4' });
   const request_ = createMessage(messageType.BINDING_REQUEST);